#101 [fix] 행동 로그 도메인 모델을 재생성하고 어댑터에서 구현할 포트를 설정한다.
