+++ build.gradle
@@ -41,35 +41,21 @@ repositories {
 }
 
 dependencies {
-	// Spring Boot
+	// Spring Core
 	implementation 'org.springframework.boot:spring-boot-starter-web'
+	implementation 'org.springframework.boot:spring-boot-starter-aop'
 	implementation 'org.springframework.boot:spring-boot-starter-validation'
 	implementation 'org.springframework.boot:spring-boot-starter-actuator'
-	implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
-
-	// JPA
 	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 
 	// MySQL
 	implementation 'com.mysql:mysql-connector-j'
 
-	// === Kafka ===
+	// Kafka
 	implementation 'org.springframework.kafka:spring-kafka'
 
-	// === Elasticsearch (Spring Data) ===
-	implementation 'org.springframework.data:spring-data-elasticsearch'
-	implementation 'co.elastic.clients:elasticsearch-java:8.13.0' // 최신 Java 클라이언트 (ES 8.x 이상용)
-
-	// === Jackson (Elasticsearch JSON 처리용) ===
-	implementation 'com.fasterxml.jackson.core:jackson-databind'
-
-	// === 로그 ===
-	implementation 'org.slf4j:slf4j-api'
-	implementation 'org.apache.logging.log4j:log4j-slf4j2-impl:2.20.0'
-
-	// Lombok
-	compileOnly 'org.projectlombok:lombok'
-	annotationProcessor 'org.projectlombok:lombok'
+	// ElasticSearch
+	implementation 'org.springframework.boot:spring-boot-starter-data-elasticsearch'
 
 	// OAuth 2.0
 	implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
@@ -78,16 +64,13 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-security'
 	testImplementation 'org.springframework.security:spring-security-test'
 
-	// Send Grid && SES & MAIL
-	implementation 'org.springframework.boot:spring-boot-starter-mail'
-	implementation 'com.sendgrid:sendgrid-java:4.9.3'
-	implementation 'com.amazonaws:aws-java-sdk-ses:1.12.3'
-
 	// Redis & Redisson
 	implementation 'org.springframework.boot:spring-boot-starter-data-redis'
 	implementation 'org.redisson:redisson:3.27.0'
-	implementation 'org.springframework.boot:spring-boot-starter-aop'
-	implementation 'org.apache.commons:commons-pool2'
+
+	// Prometheus + Actuator
+	implementation 'io.micrometer:micrometer-registry-prometheus'
+	implementation 'org.springframework.boot:spring-boot-starter-actuator'
 
 	// AWS S3
 	implementation 'com.amazonaws:aws-java-sdk-s3:1.12.643'
@@ -95,14 +78,29 @@ dependencies {
 	// Swagger
 	implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.5.0'
 
+	// JSON 처리
+	implementation 'com.fasterxml.jackson.core:jackson-databind'
+
 	// JWT → 최신 버전 사용
 	implementation 'io.jsonwebtoken:jjwt-api:0.12.3'
 	implementation 'io.jsonwebtoken:jjwt-impl:0.12.3'
 	implementation 'io.jsonwebtoken:jjwt-jackson:0.12.3'
 
-	// Prometheus
-	implementation 'io.micrometer:micrometer-registry-prometheus'
+	// MDC 로깅
+	implementation 'org.slf4j:slf4j-api'
+	runtimeOnly 'org.slf4j:jul-to-slf4j'
+	runtimeOnly 'org.slf4j:log4j-over-slf4j'
 
+	// Lombok
+	compileOnly 'org.projectlombok:lombok'
+	annotationProcessor 'org.projectlombok:lombok'
+
+	// Send Grid && SES & MAIL
+	implementation 'org.springframework.boot:spring-boot-starter-mail'
+	implementation 'com.sendgrid:sendgrid-java:4.9.3'
+	implementation 'com.amazonaws:aws-java-sdk-ses:1.12.3'
+
+	implementation 'org.apache.commons:commons-pool2'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 }
