+++ src/main/resources/application-local.yml
@@ -7,7 +7,6 @@ spring:
     import: "optional:file:.env[.properties]"
     activate:
       on-profile: local
-
   datasource:
     url: jdbc:mysql://${DB_HOST}:${DB_PORT}/${DB_NAME}?useSSL=false&allowPublicKeyRetrieval=true&serverTimezone=Asia/Seoul
     username: ${DB_USERNAME}
